+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniApp.kt
@@ -12,6 +12,7 @@ import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionCache
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
 import com.rakuten.tech.mobile.miniapp.navigator.MiniAppNavigator
 import com.rakuten.tech.mobile.miniapp.file.MiniAppFileChooser
+import com.rakuten.tech.mobile.miniapp.js.MessageBridgeRatDispatcher
 import com.rakuten.tech.mobile.miniapp.signatureverifier.SignatureVerifier
 import com.rakuten.tech.mobile.miniapp.storage.verifier.CachedMiniAppVerifier
 import com.rakuten.tech.mobile.miniapp.storage.DownloadedManifestCache
@@ -35,6 +36,14 @@ abstract class MiniApp internal constructor() {
      */
     @Throws(MiniAppSdkException::class)
     abstract suspend fun listMiniApp(): List<MiniAppInfo>
+
+    /**
+     * Fetches MiniappInfo by preview code.
+     * @return of type [MiniAppInfo] when obtained successfully
+     * @throws [MiniAppSdkException] when fetching fails from the BE server for any reason.
+     */
+    @Throws(MiniAppSdkException::class)
+    abstract suspend fun getMiniAppInfoByPreviewCode(previewCode: String): PreviewMiniAppInfo
 
     /**
      * Creates a mini app.
@@ -174,8 +183,9 @@ abstract class MiniApp internal constructor() {
 
     /**
      * Update SDK interaction interface based on [MiniAppSdkConfig] configuration.
+     * [setConfigAsDefault] for use the [MiniAppSdkConfig] as default.
      */
-    internal abstract fun updateConfiguration(newConfig: MiniAppSdkConfig)
+    internal abstract fun updateConfiguration(newConfig: MiniAppSdkConfig, setConfigAsDefault: Boolean)
 
     companion object {
         @VisibleForTesting
@@ -187,11 +197,12 @@ abstract class MiniApp internal constructor() {
          * as defined in AndroidManifest.xml. For usual scenarios the default config suffices.
          * However, should it be required to change the config at runtime for QA purpose or similar,
          * another [MiniAppSdkConfig] can be provided for customization.
+         * [setConfigAsDefault] is to use the config as default.
          * @return [MiniApp] instance
          */
         @JvmStatic
-        fun instance(settings: MiniAppSdkConfig = defaultConfig): MiniApp =
-            instance.apply { updateConfiguration(settings) }
+        fun instance(settings: MiniAppSdkConfig = defaultConfig, setConfigAsDefault: Boolean = true): MiniApp =
+            instance.apply { updateConfiguration(settings, setConfigAsDefault) }
 
         internal fun init(context: Context, miniAppSdkConfig: MiniAppSdkConfig) {
             defaultConfig = miniAppSdkConfig
@@ -199,23 +210,21 @@ abstract class MiniApp internal constructor() {
                 baseUrl = miniAppSdkConfig.baseUrl,
                 rasProjectId = miniAppSdkConfig.rasProjectId,
                 subscriptionKey = miniAppSdkConfig.subscriptionKey,
-                isPreviewMode = miniAppSdkConfig.isPreviewMode
+                isPreviewMode = miniAppSdkConfig.isPreviewMode,
+                sslPublicKey = miniAppSdkConfig.sslPinningPublicKey
             )
             val apiClientRepository = ApiClientRepository().apply {
                 registerApiClient(defaultConfig.key, apiClient)
             }
-
             val signatureVerifier: SignatureVerifier? = SignatureVerifier.init(
                 context = context,
                 baseUrl = miniAppSdkConfig.baseUrl + "keys/",
                 subscriptionKey = miniAppSdkConfig.subscriptionKey
             )
-
             val miniAppAnalytics = MiniAppAnalytics(
-                    miniAppSdkConfig.rasProjectId,
-                    miniAppSdkConfig.miniAppAnalyticsConfigList
+                miniAppSdkConfig.rasProjectId,
+                miniAppSdkConfig.miniAppAnalyticsConfigList
             )
-
             instance = RealMiniApp(
                 apiClientRepository = apiClientRepository,
                 displayer = Displayer(defaultConfig.hostAppUserAgentInfo),
@@ -233,7 +242,8 @@ abstract class MiniApp internal constructor() {
                 initCustomPermissionCache = { MiniAppCustomPermissionCache(context) },
                 initDownloadedManifestCache = { DownloadedManifestCache(context) },
                 initManifestVerifier = { MiniAppManifestVerifier(context) },
-                miniAppAnalytics = miniAppAnalytics
+                miniAppAnalytics = miniAppAnalytics,
+                ratDispatcher = MessageBridgeRatDispatcher(miniAppAnalytics = miniAppAnalytics)
             )
         }
     }