+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniappSdkInitializer.kt
@@ -61,6 +61,12 @@ class MiniappSdkInitializer : ContentProvider() {
          **/
         @MetaData(key = "com.rakuten.tech.mobile.ras.ProjectSubscriptionKey")
         fun subscriptionKey(): String
+
+        /**
+         * Public Key for the ssl pinning.
+         **/
+        @MetaData(key = "com.rakuten.tech.mobile.ras.SSLPinningPublicKey")
+        fun sslPinningPublicKey(): String
     }
 
     override fun onCreate(): Boolean {
@@ -84,7 +90,8 @@ class MiniappSdkInitializer : ContentProvider() {
         subscriptionKey = manifestConfig.subscriptionKey(),
         hostAppUserAgentInfo = manifestConfig.hostAppUserAgentInfo(),
         isPreviewMode = manifestConfig.isPreviewMode(),
-        isRequireSignatureVerification = manifestConfig.isRequireSignatureVerification()
+        isRequireSignatureVerification = manifestConfig.isRequireSignatureVerification(),
+        sslPinningPublicKey = manifestConfig.sslPinningPublicKey()
     )
 
     private fun executeMiniAppAnalytics(rasProjId: String) {