+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/api/AppInfoApi.kt
@@ -1,6 +1,7 @@
 package com.rakuten.tech.mobile.miniapp.api
 
 import com.rakuten.tech.mobile.miniapp.MiniAppInfo
+import com.rakuten.tech.mobile.miniapp.PreviewMiniAppInfo
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
@@ -20,4 +21,10 @@ internal interface AppInfoApi {
         @Path("testPath") testPath: String = "",
         @Query("miniAppId") miniAppId: String
     ): Call<List<MiniAppInfo>>
+
+    @GET("host/{hostId}/preview-codes/{previewCode}")
+    fun fetchInfoByPreviewCode(
+        @Path("hostId") hostId: String,
+        @Path("previewCode") previewCode: String = ""
+    ): Call<PreviewMiniAppInfo>
 }