+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/RealMiniAppSpec.kt
@@ -4,6 +4,7 @@ import com.rakuten.tech.mobile.miniapp.analytics.MiniAppAnalytics
 import com.rakuten.tech.mobile.miniapp.api.*
 import com.rakuten.tech.mobile.miniapp.display.Displayer
 import com.rakuten.tech.mobile.miniapp.file.MiniAppFileChooser
+import com.rakuten.tech.mobile.miniapp.js.MessageBridgeRatDispatcher
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
 import com.rakuten.tech.mobile.miniapp.navigator.MiniAppNavigator
 import com.rakuten.tech.mobile.miniapp.permission.*
@@ -36,6 +37,7 @@ open class BaseRealMiniAppSpec {
     val miniAppMessageBridge: MiniAppMessageBridge = mock()
     val miniAppNavigator: MiniAppNavigator = mock()
     val miniAppFileChooser: MiniAppFileChooser = mock()
+
     val dummyManifest: MiniAppManifest = MiniAppManifest(
         listOf(Pair(MiniAppCustomPermissionType.USER_NAME, "reason")), listOf(),
         TEST_ATP_LIST, mapOf(), TEST_MA_VERSION_ID
@@ -45,6 +47,8 @@ open class BaseRealMiniAppSpec {
         TEST_HA_ANALYTICS_CONFIGS
     )
 
+    internal var ratDispatcher = MessageBridgeRatDispatcher(miniAppAnalytics)
+
     @Before
     fun setup() {
         realMiniApp =
@@ -52,7 +56,8 @@ open class BaseRealMiniAppSpec {
                 initCustomPermissionCache = { miniAppCustomPermissionCache },
                 initDownloadedManifestCache = { downloadedManifestCache },
                 initManifestVerifier = { manifestVerifier },
-                miniAppAnalytics = miniAppAnalytics
+                miniAppAnalytics = miniAppAnalytics,
+                ratDispatcher = ratDispatcher
             ))
 
         When calling apiClientRepository.getApiClientFor(miniAppSdkConfig.key) itReturns apiClient
@@ -126,7 +131,8 @@ class RealMiniAppSpec : BaseRealMiniAppSpec() {
                 miniAppCustomPermissionCache,
                 downloadedManifestCache,
                 "",
-                miniAppAnalytics
+                miniAppAnalytics,
+                ratDispatcher
             )
         }
 
@@ -149,7 +155,8 @@ class RealMiniAppSpec : BaseRealMiniAppSpec() {
                 miniAppCustomPermissionCache,
                 downloadedManifestCache,
                 "",
-                miniAppAnalytics
+                miniAppAnalytics,
+                ratDispatcher
             )
         }
 
@@ -167,7 +174,8 @@ class RealMiniAppSpec : BaseRealMiniAppSpec() {
             verify(miniAppDownloader).validateHttpAppUrl(TEST_MA_URL)
             verify(displayer).createMiniAppDisplay(
                 TEST_MA_URL, miniAppMessageBridge, null, null,
-                miniAppCustomPermissionCache, downloadedManifestCache, "", miniAppAnalytics
+                miniAppCustomPermissionCache, downloadedManifestCache, "", miniAppAnalytics,
+                ratDispatcher
             )
         }
 
@@ -176,7 +184,7 @@ class RealMiniAppSpec : BaseRealMiniAppSpec() {
     /** region: RealMiniApp.updateConfiguration */
     @Test
     fun `should update ApiClient when configuration updated`() {
-        realMiniApp.updateConfiguration(miniAppSdkConfig)
+        realMiniApp.updateConfiguration(miniAppSdkConfig, setConfigAsDefault = true)
 
         verify(miniAppDownloader).updateApiClient(apiClient)
         verify(miniAppInfoFetcher).updateApiClient(apiClient)
@@ -186,7 +194,7 @@ class RealMiniAppSpec : BaseRealMiniAppSpec() {
     fun `should not create ApiClient for existing configuration`() {
         val miniApp = Mockito.spy(realMiniApp)
 
-        realMiniApp.updateConfiguration(miniAppSdkConfig)
+        realMiniApp.updateConfiguration(miniAppSdkConfig, setConfigAsDefault = true)
 
         verify(miniApp, times(0)).createApiClient(miniAppSdkConfig)
     }
@@ -205,7 +213,7 @@ class RealMiniAppSpec : BaseRealMiniAppSpec() {
             miniAppAnalyticsConfigList = TEST_HA_ANALYTICS_CONFIGS
         )
 
-        miniApp.updateConfiguration(miniAppSdkConfig)
+        miniApp.updateConfiguration(miniAppSdkConfig, setConfigAsDefault = true)
 
         verify(miniApp).createApiClient(miniAppSdkConfig)
     }