+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/settings/AppSettings.kt
@@ -137,7 +137,8 @@ class AppSettings private constructor(context: Context) {
                     BuildConfig.ADDITIONAL_ANALYTICS_ACC,
                     BuildConfig.ADDITIONAL_ANALYTICS_AID
                 )
-            )
+            ),
+            sslPinningPublicKey = manifestConfig.sslPinningPublicKey()
         )
 
     companion object {